Move Span into source module
