+++ fathom/src/core/binary.rs
@@ -7,8 +7,9 @@ use std::slice::SliceIndex;
 use std::sync::Arc;
 
 use crate::core::semantics::{self, ArcValue, Elim, Head, Value};
-use crate::core::{Const, Prim, Span, UIntStyle};
+use crate::core::{Const, Prim, UIntStyle};
 use crate::env::{EnvLen, SliceEnv};
+use crate::source::Span;
 
 #[derive(Clone, Debug)]
 pub enum ReadError {
