+++ fathom/src/source.rs
@@ -3,6 +3,35 @@
 /// File id.
 pub type FileId = usize; // TODO: use wrapper struct
 
+// TODO: Better name?
+#[derive(Debug, Copy, Clone)]
+pub enum Span {
+    Range(ByteRange),
+    Empty,
+}
+
+impl Span {
+    pub const fn fixme() -> Span {
+        Span::Empty
+    }
+
+    pub fn range_todo(self) {
+        // placeholder function for when we should build a span from a ByteRange
+        // but the current location is expecting a range of (). Eventually this
+        // should be renamed range and return Option<ByteRange>?
+        // match self {
+        //     Span::Range(range) => Some(range),
+        //     Span::Empty => None
+        // }
+    }
+}
+
+impl From<&ByteRange> for Span {
+    fn from(range: &ByteRange) -> Self {
+        Span::Range(*range)
+    }
+}
+
 /// Byte offsets into source files.
 pub type BytePos = usize;
